Bellabeat Google Data Analytics Capstone Code Dump
This code is for the purposes of analyzing data and making business recommendations to Bellabeat from Fitbit user data. You may run this code on your own device for reference, however you may not copy and paste and claim as your own verbatim. If this code helps you with your analysis please include this github page in your references.